Merge from origin/master

+++ hack/verify-openapi-spec.sh
@@ -14,11 +14,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# This script checks whether updating of OpenAPI specification is needed or not.
+# We should run `hack/update-openapi-spec.sh` if OpenAPI specification is out of
+# date.
+# Usage: `hack/verify-openapi-spec.sh`.
+
 set -o errexit
 set -o nounset
 set -o pipefail
 
-KUBE_ROOT=$(dirname "${BASH_SOURCE}")/..
+KUBE_ROOT=$(dirname "${BASH_SOURCE[0]}")/..
 source "${KUBE_ROOT}/hack/lib/init.sh"
 
 kube::golang::setup_env
@@ -27,15 +32,16 @@ kube::etcd::install
 make -C "${KUBE_ROOT}" WHAT=cmd/kube-apiserver
 
 SPECROOT="${KUBE_ROOT}/api/openapi-spec"
+SPECV3PATH="${SPECROOT}/v3"
 TMP_SPECROOT="${KUBE_ROOT}/_tmp/openapi-spec"
 _tmp="${KUBE_ROOT}/_tmp"
 
 mkdir -p "${_tmp}"
 cp -a "${SPECROOT}" "${TMP_SPECROOT}"
-trap "cp -a ${TMP_SPECROOT} ${SPECROOT}/..; rm -rf ${_tmp}" EXIT SIGINT
-rm ${SPECROOT}/*
-cp ${TMP_SPECROOT}/BUILD ${SPECROOT}/BUILD
-cp ${TMP_SPECROOT}/README.md ${SPECROOT}/README.md
+trap 'cp -a ${TMP_SPECROOT} ${SPECROOT}/..; rm -rf ${_tmp}' EXIT SIGINT
+rm -r "${SPECROOT:?}"/*
+mkdir -p "${SPECV3PATH}"
+cp "${TMP_SPECROOT}/README.md" "${SPECROOT}/README.md"
 
 "${KUBE_ROOT}/hack/update-openapi-spec.sh"
 echo "diffing ${SPECROOT} against freshly generated openapi spec"
